Refactor: move WPF items to separate namespaces
