+++ MediaCenter/MediaCenter/MediaCenter.csproj
@@ -106,7 +106,7 @@
       <DependentUpon>StagedItemPreviewWindow.xaml</DependentUpon>
     </Compile>
     <Compile Include="Media\PlayStateVisibilityConverter.cs" />
-    <Compile Include="WPF\BoolVisibilityConverter.cs" />
+    <Compile Include="WPF\Converters\BoolVisibilityConverter.cs" />
     <Compile Include="WPF\Controls\HideableColumnDefinition.cs" />
     <Compile Include="WPF\Controls\HideableRowDefinition.cs" />
     <Compile Include="WPF\Controls\ImageButton.cs" />
@@ -192,10 +192,10 @@
       <DependentUpon>EditStagedItemView.xaml</DependentUpon>
     </Compile>
     <Compile Include="Sessions\Staging\StagedItem.cs" />
-    <Compile Include="WPF\BindableListboxSelectionBehavior.cs" />
-    <Compile Include="WPF\NullVisibilityConverter.cs" />
-    <Compile Include="WPF\ObjectBoolConverter.cs" />
-    <Compile Include="WPF\OpenCloseWindowBehavior.cs" />
+    <Compile Include="WPF\Behaviors\BindableListboxSelectionBehavior.cs" />
+    <Compile Include="WPF\Converters\NullVisibilityConverter.cs" />
+    <Compile Include="WPF\Converters\ObjectBoolConverter.cs" />
+    <Compile Include="WPF\Behaviors\OpenCloseWindowBehavior.cs" />
     <Compile Include="Sessions\Slideshow\SlideShowViewModel.cs" />
     <Compile Include="Sessions\Slideshow\SlideShowWindow.xaml.cs">
       <DependentUpon>SlideShowWindow.xaml</DependentUpon>
@@ -215,7 +215,7 @@
       <DependentUpon>EditTagsView.xaml</DependentUpon>
     </Compile>
     <Compile Include="Media\EditTagsViewModel.cs" />
-    <Compile Include="WPF\ScrollIntoViewBehavior.cs" />
+    <Compile Include="WPF\Behaviors\ScrollIntoViewBehavior.cs" />
     <Page Include="Sessions\Filters\FilterView.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
@@ -224,7 +224,7 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="WPF\RowColumnToggleButton.xaml">
+    <Page Include="WPF\Controls\RowColumnToggleButton.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
