+++ MediaCenter/MediaCenter/Sessions/Query/QueryResultDetailView.xaml
@@ -8,6 +8,7 @@
              xmlns:media="clr-namespace:MediaCenter.Media"
              xmlns:controls="clr-namespace:MediaCenter.WPF.Controls"
              xmlns:wpf="clr-namespace:MediaCenter.WPF"
+             xmlns:behaviors="clr-namespace:MediaCenter.WPF.Behaviors"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300"
              d:DataContext="{d:DesignInstance local:QueryResultDetailViewModel}">
@@ -16,7 +17,7 @@
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
                 <ResourceDictionary Source="/Media/MediaItemViews.xaml"/>
-                <ResourceDictionary Source="/WPF/RowColumnToggleButton.xaml"/>
+                <ResourceDictionary Source="/WPF/Controls/RowColumnToggleButton.xaml"/>
             </ResourceDictionary.MergedDictionaries>
         </ResourceDictionary>
     </UserControl.Resources>
@@ -38,7 +39,7 @@
         <ListBox Grid.Row="2" Style="{StaticResource ThumbnailListStyle}" ItemsSource="{Binding QueryResultItems}" 
                  SelectedItem="{Binding Path=SelectedItem, Mode=TwoWay}" ScrollViewer.HorizontalScrollBarVisibility="Auto" >
             <interactivity:Interaction.Behaviors>
-                <wpf:ScrollIntoViewBehavior/>
+                <behaviors:ScrollIntoViewBehavior/>
             </interactivity:Interaction.Behaviors>
             <ItemsControl.ItemsPanel>
                 <ItemsPanelTemplate>
