+++ MediaCenter/MediaCenter/WPF/Behaviors/OpenCloseWindowBehavior.cs
@@ -3,7 +3,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Interactivity;
 
-namespace MediaCenter.WPF
+namespace MediaCenter.WPF.Behaviors
 {
     public class OpenCloseWindowBehavior : Behavior<UserControl>
     {
