+++ MediaCenter/MediaCenter/WPF/Behaviors/ScrollIntoViewBehavior.cs
@@ -1,7 +1,7 @@
 ﻿using System.Windows.Controls;
 using System.Windows.Interactivity;
 
-namespace MediaCenter.WPF
+namespace MediaCenter.WPF.Behaviors
 {
     public sealed class ScrollIntoViewBehavior:Behavior<ListBox>
     {
