+++ MediaCenter/MediaCenter/WPF/Converters/BoolVisibilityConverter.cs
