+++ MediaCenter/MediaCenter/WPF/Converters/NullVisibilityConverter.cs
