+++ MediaCenter/MediaCenter/WPF/Converters/ObjectBoolConverter.cs
